moved linalg submodule into core
