+++ heat/__init__.py
@@ -4,5 +4,5 @@ from . import regression
 from . import spatial
 from .core import *
 from .core import __version__
-from .linalg import *
+from .core.linalg import *
 from . import utils
