+++ heat/core/statistics.py
@@ -2,9 +2,8 @@ import numpy as np
 import torch
 
 from .communication import MPI
-from . import exponential
+from . import exponential, linalg
 from . import factories
-from .. import linalg
 from . import manipulations
 from . import operations
 from . import dndarray
